tests/wpt/mozilla/tests/webxr/obtain_frame.html

<html>
    <head>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="./resources/webxr-util.js"></script>
    </head>
    <body>
          <canvas id="canvas" width="640" height="480"></canvas>

        <script>
          let canvas = document.getElementById("canvas");
          let gl = canvas.getContext('webgl');
          promise_test(async function() {
            let mock = await navigator.xr.test.simulateDeviceConnection({supportsImmersive: true});
            mock.setViewerOrigin({position: [0.5, 0.1, 0.1, 1], orientation: [1, 0, 0, 1] });
            mock.setViews(TEST_VIEWS);
            let session = await navigator.xr.requestSession({mode: "immersive-vr"});
            await session.updateRenderState({"baseLayer": new XRWebGLLayer(session, gl, {})});
            let resolve;

            let frame = await new Promise(resolve => {
              session.requestAnimationFrame((time, frame) => resolve(frame));
            });
            
            let space = await session.requestReferenceSpace("local");
            let pose = frame.getViewerPose(space);
            for (view of pose.views) {
              assert_matrix_approx_equals(view.projectionMatrix, VALID_PROJECTION_MATRIX, 0.001, "left projection matrix");

              if (view.eye == "left") {
                assert_matrix_approx_equals(view.transform.matrix,         [1,0,0,0,0,0,1,0,0,-1,0,0,0.4,0.1,0.1,1], 0.001, "left transform");
                assert_matrix_approx_equals(view.transform.inverse.matrix, [1,0,0,0,0,0,-1,0,0,1,0,0,-0.4,-0.1,0.1,1], 0.001, "left inverse transform");
                let position = view.transform.position;
                assert_approx_equals(position.x, 0.4, 0.001, "left x value");
                assert_approx_equals(position.y, 0.1, 0.001, "left y value");
                assert_approx_equals(position.z, 0.1, 0.001, "left z value");
                assert_approx_equals(position.w, 1, 0.001, "left w value");
              } else if (view.eye == "right") {
                assert_matrix_approx_equals(view.transform.matrix,         [1,0,0,0,0,0,1,0,0,-1,0,0,0.6,0.1,0.1,1], 0.001, "right transform");
                assert_matrix_approx_equals(view.transform.inverse.matrix, [1,0,0,0,0,0,-1,0,0,1,0,0,-0.6,-0.1,0.1,1], 0.001, "right inverse transform");
                let position = view.transform.position;
                assert_approx_equals(position.x, 0.6, 0.001, "left x value");
                assert_approx_equals(position.y, 0.1, 0.001, "left y value");
                assert_approx_equals(position.z, 0.1, 0.001, "left z value");
                assert_approx_equals(position.w, 1, 0.001, "left w value");
              } else {
                reject("got unknown view");
              }
            }

          });
        </script>
    </body>
</html>
